# Conflicts:
#	crest/Assets/Crest/Crest/Shaders/Underwater/UnderwaterCurtain.shader
#	crest/Assets/Crest/Crest/Shaders/Underwater/UnderwaterMeniscus.shader

+++ crest/Assets/Crest/Crest/Shaders/OceanInputs/Resources/OceanDepthsCache.shader
@@ -53,7 +53,7 @@ Shader "Crest/Inputs/Depth/Cached Depths"
 
 			float2 Frag(Varyings input) : SV_Target
 			{
-				return half2(tex2D(_MainTex, input.uv).x, 0.0);
+				return float2(tex2D(_MainTex, input.uv).x, 0.0);
 			}
 			ENDCG
 		}
